// new/sigbaux-front/src/app/interfaces/ministerial-structure.ts
export interface MinisterialStructure {
    id: number | null;
    name: string;
    domain: string;
    phone: string;
    email: string;
    manager: string | null;
    code: string;
    idMinistry: number;
}

/*
export interface MinisterialStructure {
    id?: number;
    name?: string;
    domain?: string;
    phone?: string;
    email?: string;
    idMinistry?: string;
    manager?: string;
    code?: string;
}
*/